+++ accounts-db/src/lib.rs
@@ -13,6 +13,7 @@
 pub mod account_info;
 pub mod account_locks;
 pub mod account_storage;
+pub mod account_storage_entry;
 pub mod account_storage_reader;
 pub mod accounts;
 mod accounts_cache;
