+++ accounts-db/src/storable_accounts.rs
@@ -2,7 +2,8 @@
 use {
     crate::{
         account_storage::stored_account_info::StoredAccountInfo,
-        accounts_db::{AccountFromStorage, AccountStorageEntry, AccountsDb},
+        account_storage_entry::AccountStorageEntry,
+        accounts_db::{AccountFromStorage, AccountsDb},
         is_zero_lamport::IsZeroLamport,
         utils::create_account_shared_data,
     },
@@ -363,7 +364,8 @@ pub mod tests {
         super::*,
         crate::{
             account_info::{AccountInfo, StorageLocation},
-            accounts_db::{get_temp_accounts_paths, AccountStorageEntry},
+            account_storage_entry::AccountStorageEntry,
+            accounts_db::get_temp_accounts_paths,
             accounts_file::AccountsFileProvider,
         },
         rand::Rng,
