+++ snapshots/src/archive.rs
@@ -6,8 +6,8 @@ use {
     agave_fs::buffered_writer::large_file_buf_writer,
     log::info,
     solana_accounts_db::{
-        account_storage::AccountStoragesOrderer, account_storage_reader::AccountStorageReader,
-        accounts_db::AccountStorageEntry, accounts_file::AccountsFile,
+        account_storage::AccountStoragesOrderer, account_storage_entry::AccountStorageEntry,
+        account_storage_reader::AccountStorageReader, accounts_file::AccountsFile,
     },
     solana_clock::Slot,
     solana_measure::measure::Measure,
